+++ R/locate_credentials.R
@@ -11,14 +11,25 @@
 #' @param verbose A logical indicating whether to be verbose.
 #' @details These functions locate values of AWS credentials (access key, secret access key, session token, and region) from likely sources. The order in which these are searched is as follows:
 #' \enumerate{
-#'   \item values passed to the function
-#'   \item environment variables (\env{AWS_ACCESS_KEY_ID}, \env{AWS_SECRET_ACCESS_KEY}, \env{AWS_SESSION_TOKEN}, \env{AWS_DEFAULT_REGION})
+#'   \item user-supplied values passed to the function
+#'   \item environment variables (\env{AWS_ACCESS_KEY_ID}, \env{AWS_SECRET_ACCESS_KEY}, \env{AWS_SESSION_TOKEN})
 #'   \item an IAM instance role (on the running EC2 instance from which this function is called) as identified by \code{\link[aws.ec2metadata]{metadata}}
 #'   \item a specified profile in a local credentials dot file in the current working directory
 #'   \item the default profile in a local credentials dot file in the current working directory
 #'   \item a specified profile in a global credentials dot file in, typically in \file{~/.aws/credentials}. See \code{\link{use_credentials}} for details
 #'   \item the default profile in a global credentials dot file in, typically in \file{~/.aws/credentials}. See \code{\link{use_credentials}} for details
 #' }
+#' 
+#' Because region is often handled slightly differently from credentials and is required for most requests (whereas some services allow anonymous requests without specifying credentials), the value of region is searched for in the same order as the above but lacking a value there fails safe with the following preference ranking of possible region values (regardless of location of other credentials):
+#' \enumerate{
+#'   \item a user-supplied value
+#'   \item (if a credentials file is being used) the value specified therein
+#'   \item the \env{AWS_DEFAULT_REGION} environment variable
+#'   \item (only on EC2 instances) a region declared in the instance metadata
+#'   \item the default value specified in \code{default_region} (i.e., \dQuote{us-east-1})
+#' }
+#' 
+#' As such, user-supplied values of \code{region} always trump any other value.
 #' 
 #' @seealso \code{\link{signature_v4}}, \code{\link{signature_v2_auth}}, \code{\link{use_credentials}}
 #' @export
@@ -55,9 +66,15 @@ function(key = NULL,
                 message("Using user-supplied value for AWS Session Token")
             }
         }
-        if (!is.null(region)) {
+        # now find region, with fail safes
+        if (!is.null(region) && region != "") {
             if (isTRUE(verbose)) {
                 message(sprintf("Using user-supplied value for AWS Region ('%s')", region))
+            }
+        } else if (!is.null(env$region) && env$region != "") {
+            region <- env$region
+            if (isTRUE(verbose)) {
+                message(sprintf("Using Environment Variable 'AWS_DEFAULT_REGION' for AWS Region ('%s')", region))
             }
         } else {
             region <- default_region
@@ -66,60 +83,65 @@ function(key = NULL,
             }
         }
     } else if ((!is.null(env$key) && env$key != "") || (!is.null(env$secret) && env$secret != "")) {
+        # otherwise use environment variables if no user-supplied values
         if (!is.null(env$key) && env$key != "") {
+            key <- env$key
             if (isTRUE(verbose)) {
                 message("Using Environment Variable 'AWS_ACCESS_KEY_ID' for AWS Access Key ID")
             }
         } else {
             if (!is.null(key) && key != "") {
-                env$key <- key
+                key <- key
                 if (isTRUE(verbose)) {
                     message("Using user-supplied value for AWS Access Key ID")
                 }
             }
         }
         if (!is.null(env$secret) && env$secret != "") {
+            secret <- env$secret
             if (isTRUE(verbose)) {
                 message("Using Environment Variable 'AWS_SECRET_ACCESS_KEY' for AWS Secret Access Key")
             }
         } else {
             if (!is.null(secret) && secret != "") {
-                env$secret <- secret
+                secret <- secret
                 if (isTRUE(verbose)) {
                     message("Using user-supplied value for AWS Secret Access Key")
                 }
             }
         }
         if (!is.null(env$session_token) && env$session_token != "") {
+            session_token <- env$session_token
             if (isTRUE(verbose)) {
                 message("Using Environment Variable 'AWS_SESSION_TOKEN' for AWS Session Token")
             }
         } else {
             if (!is.null(session_token) && session_token != "") {
-                env$session_token <- session_token
+                session_token <- session_token
                 if (isTRUE(verbose)) {
                     message("Using user-supplied value for AWS Session Token")
                 }
             }
         }
-        if (!is.null(region) && region != "") {
-            env$region <- region
+        # now find region, with fail safes
+        if (!is.null(region)) {
+            region <- region
             if (isTRUE(verbose)) {
                 message(sprintf("Using user-supplied value for AWS Region ('%s')", region))
             }
-        } else if (is.null(env$region) || env$region == "") {
-            env$region <- default_region
+        } else if (!is.null(env$region) && env$region != "") {
+            region <- env$region
+            if (isTRUE(verbose)) {
+                message(sprintf("Using Environment Variable 'AWS_DEFAULT_REGION' for AWS Region ('%s')", region))
+            }
+        } else {
+            region <- default_region
             if (isTRUE(verbose)) {
                 message(sprintf("Using default value for AWS Region ('%s')", region))
             }
-        } else {
-            if (isTRUE(verbose)) {
-                message("Using Environment Variable 'AWS_DEFAULT_REGION' for AWS Region")
-            }
         }
-        return(env)
     } else {
-        # check for EC2 metadata
+        # lacking that, check for EC2 metadata
         role <- try(get_ec2_role(verbose = verbose), silent = TRUE)
         if (!inherits(role, "try-error")) {
             if (!is.null(role[["AccessKeyId"]])) {
@@ -140,19 +162,33 @@ function(key = NULL,
                     message("Using EC2 Instance Metadata for AWS Session Token")
                 }
             }
-            if (!is.null(region)) {
-                region <- region
+            # now find region, with fail safes
+            if (!is.null(region) && region != "") {
                 if (isTRUE(verbose)) {
                     message(sprintf("Using user-supplied value for AWS Region ('%s')", region))
                 }
-            } else {
-                region <- default_region
+            } else if (!is.null(env$region) && env$region != "") {
+                region <- env$region
                 if (isTRUE(verbose)) {
-                    message(sprintf("Using default value for AWS Region ('%s')", region))
+                    message(sprintf("Using Environment Variable 'AWS_DEFAULT_REGION' for AWS Region ('%s')", region))
+                }
+            } else {
+                # check instance metadata for region
+                reg <- try(aws.ec2metadata::instance_document()$region, silent = TRUE)
+                if (!inherits(reg, "try-error") && !is.null(reg) && reg != "") {
+                    region <- reg
+                    if (isTRUE(verbose)) {
+                        message(sprintf("Using EC2 Instance Metadata for AWS Region ('%s')", region))
+                    }
+                } else {
+                    region <- default_region
+                    if (isTRUE(verbose)) {
+                        message(sprintf("Using default value for AWS Region ('%s')", region))
+                    }
                 }
             }
         } else {
-            # check for credentials file
+            # lastly, check for credentials file
             if (file.exists(file.path(".aws", "credentials"))) {
                 ## in working directory
                 cred <- read_credentials(file.path(".aws", "credentials"))[[profile]]
@@ -166,38 +202,45 @@ function(key = NULL,
                     message(sprintf("Using global credentials files from '%s'", default_credentials_file()))
                 }
             } else {
+                # if that fails, no credentials can be found anywhere
                 if (isTRUE(verbose)) {
                     message("No instance metadata, environment variables, or credentials file found!")
                 }
+                # return early with list of empty values!
                 return(list(key = NULL, secret = NULL, session_token = NULL, region = NULL))
             }
             if (!is.null(cred[["AWS_ACCESS_KEY_ID"]])) {
                 key <- cred[["AWS_ACCESS_KEY_ID"]]
                 if (isTRUE(verbose)) {
-                    message("Using EC2 Instance Metadata for AWS Access Key ID")
+                    message("Using value in credentials file for AWS Access Key ID")
                 }
             }
             if (!is.null(cred[["AWS_SECRET_ACCESS_KEY"]])) {
                 secret <- cred[["AWS_SECRET_ACCESS_KEY"]]
                 if (isTRUE(verbose)) {
-                    message("Using EC2 Instance Metadata for AWS Secret Access Key")
+                    message("Using value in credentials file for AWS Secret Access Key")
                 }
             }
             if (!is.null(cred[["AWS_SESSION_TOKEN"]])) {
                 session_token <- cred[["AWS_SESSION_TOKEN"]]
                 if (isTRUE(verbose)) {
-                    message("Using EC2 Instance Metadata for AWS Session Token")
+                    message("Using value in credentials file for AWS Session Token")
                 }
             }
-            if (!is.null(cred[["AWS_DEFAULT_REGION"]])) {
-                region <- cred[["AWS_DEFAULT_REGION"]]
-                if (isTRUE(verbose)) {
-                    message(sprintf("Using EC2 Instance Metadata for AWS Region ('%s')", region))
-                }
-            } else if (!is.null(region)) {
-                region <- region
+            # now find region, with fail safes
+            if (!is.null(region) && region != "") {
                 if (isTRUE(verbose)) {
                     message(sprintf("Using user-supplied value for AWS Region ('%s')", region))
+                }
+            } else if (!is.null(cred[["AWS_DEFAULT_REGION"]]) && cred[["AWS_DEFAULT_REGION"]] != "") {
+                region <- cred[["AWS_DEFAULT_REGION"]]
+                if (isTRUE(verbose)) {
+                    message(sprintf("Using value in credentials file for AWS Region ('%s')", region))
+                }
+            } else if (!is.null(env$region) && env$region != "") {
+                region <- env$region
+                if (isTRUE(verbose)) {
+                    message(sprintf("Using Environment Variable 'AWS_DEFAULT_REGION' for AWS Region ('%s')", region))
                 }
             } else {
                 region <- default_region
@@ -207,6 +250,7 @@ function(key = NULL,
             }
         }
     }
+    # return identified values
     list(key = key, secret = secret, session_token = session_token, region = region)
 }