+++ NEWS.md
@@ -1,3 +1,8 @@
+# CHANGES TO aws.signature 0.3.6
+
+* `locate_credentials()` now attempts to look in instance metadata for a region, when called from an EC2 instance. (see https://github.com/cloudyr/aws.s3/issues/151)
+* The set of fall backs for values of `region` have been standardized and documented for `locate_credentials()`.
+
 # CHANGES TO aws.signature 0.3.5
 
 * `signature_v4_auth()` and `signature_v2_auth()` now both return a `Region` value in their response list, as identified by `locate_credentials()`.